requirements.txt
# Snowflake GDPR Right to be Forgotten Demo - Python Dependencies
# Requires Python 3.9 or newer

# Snowflake connector (optimized for Python 3.9+)
snowflake-connector-python>=3.7.0,<4.0.0

# Data manipulation and analysis
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Web framework (for simple demo interface)
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0

# Configuration management
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.1,<7.0.0

# Logging and monitoring
structlog>=23.2.0,<24.0.0

# Utilities
click>=8.1.0,<9.0.0
tabulate>=0.9.0,<1.0.0
colorama>=0.4.6,<1.0.0

# Development and testing
pytest>=7.4.0,<8.0.0
black>=23.11.0,<24.0.0
flake8>=6.1.0,<7.0.0

# Optional: For more advanced data visualization
plotly>=5.17.0,<6.0.0
dash>=2.14.0,<3.0.0

# Optional: For Excel export of compliance reports
openpyxl>=3.1.0,<4.0.0

